refactor: #23 Renaming Utils -> HttpClientUtils
