+++ cogni-libs/src/main/java/zone/cogni/libs/sparqlservice/impl/StardogSparqlService.java
@@ -1,7 +1,7 @@
 package zone.cogni.libs.sparqlservice.impl;
 
 import static com.google.common.net.HttpHeaders.CONTENT_TYPE;
-import static zone.cogni.libs.sparqlservice.impl.Utils.execute;
+import static zone.cogni.libs.sparqlservice.impl.HttpClientUtils.execute;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -29,7 +29,7 @@ public class StardogSparqlService implements SparqlService {
 
   public StardogSparqlService(Config config) {
     endpointUrl = config.getUrl();
-    httpClient = Utils.createHttpClientBuilder(config.getUser(), config.getPassword()).build();
+    httpClient = HttpClientUtils.createHttpClientBuilder(config.getUser(), config.getPassword()).build();
     queryExecutionBuilder = QueryExecutionHTTPBuilder.service(endpointUrl + "/query").httpClient(httpClient);
   }
